Update JSONField to implement DjangoJSONEncoder
This commit steals some code from the head of the Django master (1.11?) to
allow the use of a custom JSONEncoder as part of the JSONField
serialization process. This means that you can pass dates and decimals
into the query and hits properties of a search query. NB when you retrieve
a search from the database, they will come back out as strings!