Fix Model._default_manager bug in 1.10
The model._default_manager attr doesn't work in 1.10
so reverting to using model.objects instead. This does
mean that if there are multiple model managers configured
this will cause a change in behaviour if the default manager
was not assigned to `objects`. However, since this is the
recommended pattern, I am assuming this is how most
applications are configured, and we will cross that
bridge when we come to it.