Update master branch to ES6 only (#36)
This commit will update the master branch to require ES6 and above. It formally deprecates ES5 and ES2, which will now live on in named branches only. Aside from documentation, and updating the dependency versions, this commit also create a separate code path for partial document updates, so that instead of a single as_search_document method that takes an optional update_fields kwarg, we have two separate methods, as_search_document and as_search_document_update. Only the later has the update_fields arg, which is mandatory. There are examples in the README, so recommendation is to start there.

One other addition is that the repo has been "Blackened" - so the formatting of app *.py files should now be consistent.