Add support for custom primary key (#58)
This change is needed to handle update index operation and lookup in search queryset.

Prior to this change, if a model had custom primary key this approach fails with an AttributeError because django doesn't create an id column for that model.

This change modifies the lookup function to search using the primary key. For regular models, django auto creates an id column and primary key returns the id, for models with a custom primary key, instance.pk returns the correct field.